api_yamdb/requirements.txt
gunicorn==20.0.4
psycopg2-binary==2.8.6
requests==2.26.0
django==2.2.28
djangorestframework==3.12.4
PyJWT==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==21.1
djangorestframework-simplejwt==4.7.2
